day02: add solution 2
The strategy book changed. The input is not longer a simple game plan, but a
more involved set of instructions.

Instead of X, Y and Z meaning "play Rock, Paper or Scissor", it became "lose,
draw or win".

So I opted to first translate the input according to the new strat (meaning
that "A X" becomes "A Z", etc) and then proceed to use the same function as
the one I used in the first solution.